JavaBox version with promises. Code is SID, untested and unchecked.
